Resolve "Remove implicit casts of integers and floating-point values"

Closes #301

See merge request MusicScience37Projects/numerical-analysis/numerical-collection-cpp!326

+++ include/num_collect/util/static_stack.h
@@ -109,7 +109,7 @@ public:
      * \return Current number of values.
      */
     [[nodiscard]] auto size() const noexcept -> std::size_t {
-        return end_ - begin_;
+        return static_cast<std::size_t>(end_ - begin_);
     }
 
 private:
